requirements.txt
django==2.2.2
djangorestframework==3.9.4
django-extensions==2.1.9
django-rest-swagger==2.2.0
django-rest-auth==0.9.5
django-allauth==0.39.1
django-rest-swagger==2.1.2
ipython==7.5.0
psycopg2-binary==2.8.3


